@@ -55,6 +55,6 @@
 		"postcss": "^8.4.16",
 		"prisma": "4.1.0",
 		"tailwindcss": "^3.1.8",
-		"typescript": "4.8.3"
+		"typescript": "5.5.4"
 	}
 }
