@@ -1,15 +1,22 @@
 {
 	"name": "@ideaslab/db",
 	"version": "1.0.0",
-	"main": "./src/index.ts",
+	"main": "./dist/index.js",
+	"module": "./dist/index.mjs",
+	"typings": "./dist/index.d.ts",
+	"exports": {
+		"import": "./dist/index.mjs",
+		"require": "./dist/index.js",
+		"types": "./dist/index.d.ts"
+	},
 	"files": [
 		"dist/**"
 	],
 	"license": "MIT",
 	"scripts": {
-		"build": "tsup ./src/index.ts --format esm,cjs",
+		"build": "tsup ./src/index.ts",
 		"lint": "TIMING=1 eslint \"src/**/*.ts\"",
-		"dev": "tsup ./src/index.ts --format esm,cjs --watch",
+		"dev": "tsup ./src/index.ts --watch",
 		"format": "prisma format",
 		"studio": "prisma studio",
 		"db:migrate:deploy": "prisma migrate deploy",
@@ -19,9 +26,10 @@
 	},
 	"devDependencies": {
 		"@ideaslab/eslint-config": "*",
+		"eslint": "8.23.1",
 		"prisma": "^4.8.1",
 		"tsup": "^6.5.0",
-		"typescript": "^4.9.4"
+		"typescript": "^5.5.4"
 	},
 	"dependencies": {
 		"@prisma/client": "^4.8.1"
