@@ -1,7 +1,7 @@
 import { initTRPC } from '@trpc/server'
 import superjson from 'superjson'
 
-import { Context } from './context'
+import { Context } from './context.js'
 
 const t = initTRPC.context<Context>().create({
   transformer: superjson,
